+++ lib/minitest/capybara.rb
@@ -21,17 +21,3 @@ end
 # Need to be required after Minitest::Capybara is defined
 require "capybara/assertions"
 require "capybara/expectations"
-
-# :stopdoc:
-module Minitest
-  module Capybara
-    def self.const_missing const
-      if :Assertions == const
-        warn "Minitest::Capybara::Assertions is deprecated. Please use Capybara::Assertions instead."
-        ::Capybara::Assertions
-      else
-        super const
-      end
-    end
-  end
-end