+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+* Remove depracation for Minitest::Capybara::Assertions
 * Better error messages for `assert_content/refute_content`.
 
 ## v0.7.2